Time: 0 ms (100.00%), Space: 5.8 MB (73.49%) - LeetHub

+++ 0371-sum-of-two-integers/0371-sum-of-two-integers.cpp
@@ -8,8 +8,8 @@ public:
         
         for(int i = 0;i<=31;i++){
             
-            int bitA = (a>>i) & 1; 
-            int bitB = (b>>i) & 1;
+            short bitA = (a>>i) & 1; 
+            short bitB = (b>>i) & 1;
 
             
             ans = ans | ((bitA ^ bitB ^ lastCarry)<<i) ;
